Move FailureZone to own file
